requirements.txt
Flask==2.2.2
opencv-python==4.6.0.66
numpy==1.23.5
pyserial==3.5
pigpio==1.78
setuptools